# Optimize .+ regex line filter