@MasslessParticle 2023-03-01
@jeschkies @owen-d Looks like it improved the performance pretty substantially:
```
name                               old time/op    new time/op    delta
_LineFilter/simplified_true_.+-8     45.2ns ± 2%     2.0ns ± 7%  -95.56%  (p=0.000 n=8+10)
_LineFilter/simplified_false_.+-8    50.4ns ±16%     3.0ns ± 2%  -94.05%  (p=0.000 n=10+8)

name                               old alloc/op   new alloc/op   delta
_LineFilter/simplified_true_.+-8      0.00B          0.00B          ~     (all equal)
_LineFilter/simplified_false_.+-8     0.00B          0.00B          ~     (all equal)

name                               old allocs/op  new allocs/op  delta
_LineFilter/simplified_true_.+-8       0.00           0.00          ~     (all equal)
_LineFilter/simplified_false_.+-8      0.00           0.00          ~     (all equal)
```